Commonize system descriptor API, update metal APIs
