+++ runtime/include/tt/runtime/detail/ttmetal.h
@@ -39,6 +39,7 @@
 #pragma clang diagnostic ignored "-Wlogical-op-parentheses"
 #pragma clang diagnostic ignored "-Wundefined-inline"
 #define FMT_HEADER_ONLY
+#include "impl/device/device_mesh.hpp"
 #include "tt_metal/host_api.hpp"
 #pragma clang diagnostic pop
